requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
WTForms
email_validator
Werkzeug
pytz
python-dateutil
gunicorn